Move generator operator tests to `generators-api`
