+++ generators-api/src/commonTest/kotlin/com/github/jcornaz/kwik/generator/FilterTest.kt
@@ -1,7 +1,6 @@
-package com.github.jcornaz.kwik.operator
+package com.github.jcornaz.kwik.generator
 
 import com.github.jcornaz.kwik.*
-import com.github.jcornaz.kwik.generator.Generator
 import com.github.jcornaz.kwik.generator.test.AbstractGeneratorTest
 import kotlin.test.Test
 import kotlin.test.assertEquals
