+++ generators-api/src/commonTest/kotlin/com/github/jcornaz/kwik/generator/PlusOperatorTest.kt
@@ -1,8 +1,6 @@
-package com.github.jcornaz.kwik.operator
+package com.github.jcornaz.kwik.generator
 
 import com.github.jcornaz.kwik.generator.test.AbstractGeneratorTest
-import com.github.jcornaz.kwik.generator.Generator
-import com.github.jcornaz.kwik.generator.plus
 import com.github.jcornaz.kwik.withSamples
 import kotlin.test.Test
 import kotlin.test.assertEquals
