+++ generators-api/src/commonTest/kotlin/com/github/jcornaz/kwik/generator/WithSampleTest.kt
@@ -1,7 +1,6 @@
-package com.github.jcornaz.kwik.operator
+package com.github.jcornaz.kwik.generator
 
 import com.github.jcornaz.kwik.generator.test.AbstractGeneratorTest
-import com.github.jcornaz.kwik.generator.Generator
 import com.github.jcornaz.kwik.withNull
 import com.github.jcornaz.kwik.withSamples
 import kotlin.test.Test
